fuentes/sea/pyproject.toml
[project]
name = "sea-seia"
version = "0.1.0"
description = "Extractor de datos del Sistema de Evaluación de Impacto Ambiental (SEA) de Chile"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mysql-connector-python>=9.4.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]
